gift: use built-in benchmarks instead of Criterion
