+++ gift/Cargo.toml
@@ -17,18 +17,9 @@ cipher = "=0.5.0-pre.6"
 
 [dev-dependencies]
 cipher = { version = "=0.5.0-pre.6", features = ["dev"] }
-criterion = "0.3.5"
-
-[target.'cfg(any(target_arch = "x86", target_arch = "x86_64"))'.dev-dependencies]
-criterion-cycles-per-byte = "0.1.2"
 
 [features]
 zeroize = ["cipher/zeroize"]
 
 [package.metadata.docs.rs]
 rustdoc-args = ["--cfg", "docsrs"]
-
-[[bench]]
-name = "gift128enc"
-path = "benches/gift128enc.rs"
-harness = false
