+++ gift/benches/gift128enc.rs
@@ -1,32 +1,25 @@
-use criterion::{criterion_group, criterion_main, BenchmarkId, Criterion, Throughput};
-use criterion_cycles_per_byte::CyclesPerByte;
-use gift_cipher::cipher::{BlockCipherEncrypt, KeyInit};
+#![feature(test)]
+extern crate test;
+
+use cipher::{block_decryptor_bench, block_encryptor_bench, KeyInit};
 use gift_cipher::Gift128;
 
-const KB: usize = 1024;
-
-fn bench(c: &mut Criterion<CyclesPerByte>) {
-    let mut group = c.benchmark_group("gift128enc");
-
-    let cipher128 = Gift128::new(&Default::default());
-
-    for size in &[KB, 2 * KB, 4 * KB, 8 * KB, 16 * KB] {
-        let mut buf = vec![Default::default(); *size / 16];
-
-        group.throughput(Throughput::Bytes(*size as u64));
-
-        group.bench_function(BenchmarkId::new("encrypt-128", size), |b| {
-            b.iter(|| cipher128.encrypt_blocks(&mut buf));
-        });
-    }
-
-    group.finish();
-}
-
-criterion_group!(
-    name = benches;
-    config = Criterion::default().with_measurement(CyclesPerByte);
-    targets = bench
+block_encryptor_bench!(
+    Key: Gift128,
+    gift128_encrypt_block,
+    gift128_encrypt_blocks,
+);
+block_decryptor_bench!(
+    Key: Gift128,
+    gift128_decrypt_block,
+    gift128_decrypt_blocks,
 );
 
-criterion_main!(benches);
+#[bench]
+fn gift128_new(bh: &mut test::Bencher) {
+    bh.iter(|| {
+        let key = test::black_box(Default::default());
+        let cipher = Gift128::new(&key);
+        test::black_box(&cipher);
+    });
+}
